fix: update all configs to latest SDK versions, remove unused deps, fix type errors
Android:
- compileSdk/targetSdk pinned to 35 (latest stable)
- minSdk set to 24 (required by camera/speech plugins)
- Kotlin bumped to 2.0.21 (compatible with AGP 8.7)
- Added multiDexEnabled

iOS:
- Added Podfile with platform :ios, '16.0'
- Added AppDelegate.swift with FlutterActivity v2 registration
- Added AppFrameworkInfo.plist with MinimumOSVersion 16.0

pubspec.yaml:
- Removed unused deps causing plugin warnings: lottie, sqflite,
  video_player, share_plus
- Bumped all remaining deps to latest compatible versions

Dart fixes:
- Fixed all .clamp() calls to use double literals (0.0, 1.0) instead
  of int literals — prevents num-to-double type errors at compile time

+++ lib/services/speech_analysis_service.dart
@@ -140,12 +140,12 @@ class SpeechAnalysisService {
     double score = 100;
 
     // Penalize for filler words (each filler word = -1.5 points, capped at -30)
-    final fillerPenalty = (fillerWordCount * 1.5).clamp(0, 30).toDouble();
+    final fillerPenalty = (fillerWordCount * 1.5).clamp(0.0, 30.0);
     score -= fillerPenalty;
 
     // Penalize for excessive pauses (>5 long pauses = penalty)
     if (pauseCount > 5) {
-      score -= ((pauseCount - 5) * 2.0).clamp(0, 20);
+      score -= ((pauseCount - 5) * 2.0).clamp(0.0, 20.0);
     }
 
     // Penalize for very short responses (under 30 seconds indicates low confidence)
@@ -165,7 +165,7 @@ class SpeechAnalysisService {
       }
     }
 
-    return score.clamp(0, 100);
+    return score.clamp(0.0, 100.0);
   }
 
   double _calculateClarityScore({
@@ -192,7 +192,7 @@ class SpeechAnalysisService {
       }
     }
 
-    return score.clamp(0, 100);
+    return score.clamp(0.0, 100.0);
   }
 
   double _calculatePaceScore(double wordsPerMinute) {
